Fix optional feature feed items
